+++ CHANGES.md
@@ -1,6 +1,6 @@
 ## vx.x.x - xxdatexx
 
-* **FEATURE** detector lookup class added alongside yaml files to host detector specific parameters
+* **FEATURE** detector lookup class added alongside yaml files to host detector specific parameters (rotation, science-pixels etc). Code has been updated to remove hard-wired detector values.
 * inject a 'SOXSPIPE PRE' keyword with timestamp value into prepared frames
 * check frames for 'SOXSPIPE PRE' keyword before preparing - raises exception if found
 * moved basic input frame verifications to the `_base_recipe` - so not to repeat code